feat: implement workspace-scoped lead capture integrations with configurable embed settings and migration support

+++ packages/Webkul/Lead/src/Database/Migrations/2026_08_13_000003_create_lead_capture_logs_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->unsignedBigInteger('connector_id')->nullable();
             $table->json('raw_payload')->nullable();
             $table->string('status')->default('success'); // success, duplicate_flagged, error
-            $table->unsignedBigInteger('lead_id')->nullable();
+            $table->unsignedInteger('lead_id')->nullable();
             $table->text('error_message')->nullable();
             $table->timestamps();
